+++ GameDetails/GameView.qml
@@ -35,7 +35,12 @@ FocusScope {
     property string collectionName: game ? game.collections.get(0).name : ""
     property string collectionShortName: game ? game.collections.get(0).shortName : ""
     property bool iamsteam: game ? (collectionShortName === "steam") : false
-    property bool canPlayVideo: settings.VideoPreview === "Yes"
+    property bool canPlayVideo: ((settings.VideoPreview === "Yes") && (appWindow.activeFocusItem !== null)) ? true : false
+    onCanPlayVideoChanged:{
+        //to force to stop video playing when we lost focus
+       if(canPlayVideo === false) toggleVideo(false);
+    }
+
     property real detailsOpacity: ((settings.DetailsDefault === "Yes") && (demoLaunched !== true)) || ((settings.DemoShowFullDetails === "Yes") && (demoLaunched === true)) ? 1 : 0
 	property real retroachievementsOpacity: 0
 	property bool blurBG: settings.GameBlurBackground === "Yes"
@@ -827,6 +832,99 @@ FocusScope {
             height: parent.height
             selected: (demoLaunched !== true) && ListView.isCurrentItem && menu.focus
             onHighlighted: { menu.currentIndex = ObjectModel.index; content.currentIndex = 0; }
+            property var launchedGame: api.launchedgame
+            property var selectedGame : game
+            property bool canLoadSpinner: (appWindow.activeFocusItem !== null) ? true : false
+
+            //Done to avoid spinner running during gaming ;-)
+            onCanLoadSpinnerChanged:{
+                if (launchedGame) {
+                    if (api.launchedgame.path === game.path){
+                        //if can't Load Spinner
+                        if(!canLoadSpinner){
+                            //stop  animation (existing or not)
+                            iconRotation.stop();
+                        }
+                        else {
+                            //restart or start animation
+                            icon = "../assets/images/loading.png";
+                            //start animation without end
+                            iconRotation.loops = Animation.Infinite;
+                            iconRotation.start();
+                        }
+                    }
+                }
+            }
+
+            onSelectedGameChanged: {
+                //console.log("onSelectedGameChanged");
+                if (launchedGame) {
+                    if (api.launchedgame.path === game.path){
+                        icon = "../assets/images/loading.png";
+                        //start animation without end
+                        iconRotation.loops = Animation.Infinite;
+                        iconRotation.start();
+                    }
+                    else{
+                        //if animation is running
+                        if(iconRotation.running === true){
+                            //stop existing animation
+                            iconRotation.stop();
+                            //limit to one cycle
+                            iconRotation.loops = 1;
+                            //restart animation
+                            iconRotation.start();
+                        }
+                        else icon = "../assets/images/icon_play.svg";
+                    }
+                }
+                else icon = "../assets/images/icon_play.svg";
+            }
+            iconRotation.onStarted: {
+                    //console.log("iconRotation.onStarted");
+                    icon = "../assets/images/loading.png";
+            }
+
+            iconRotation.onFinished:{
+                    //console.log("iconRotation.onFinished");
+                    //if finish and during the last cycle
+                    if (iconRotation.loops === 1) icon = "../assets/images/icon_play.svg";
+            }
+
+            onLaunchedGameChanged: {
+                //if game is launched
+                if(launchedGame){
+                    //if launched game is the selected game of gameview
+                    if((launchedGame.path === selectedGame.path) && (!iconRotation.running)){
+                        //start animation without end
+                        iconRotation.loops = Animation.Infinite;
+                        iconRotation.start();
+                    }
+                    else{
+                        //if animation is running
+                        if(iconRotation.running === true){
+                            //stop existing animation
+                            iconRotation.stop();
+                            //limit to one cycle
+                            iconRotation.loops = 1;
+                            //restart animation
+                            iconRotation.start();
+                        }
+                    }
+                }
+                else{
+                    //if animation is running
+                    if(iconRotation.running === true){
+                        //stop existing animation
+                        iconRotation.stop();
+                        //limit to one cycle
+                        iconRotation.loops = 1;
+                        //restart animation
+                        iconRotation.start();
+                    }
+                }
+            }
+
             onActivated:
                 if (selected) {
                     sfxAccept.play();
@@ -1119,7 +1217,7 @@ FocusScope {
         }
 
         // Next game
-        if (api.keys.isNextPage(event) && !event.isAutoRepeat) {
+        if (api.keys.isNextPage(event) && !event.isAutoRepeat && !global.guideButtonPressed) {
             event.accepted = true;
             sfxToggle.play();
             if (currentGameIndex < game.collections.get(0).games.count-1)
@@ -1131,7 +1229,7 @@ FocusScope {
         }
 
         // Previous game
-        if (api.keys.isPrevPage(event) && !event.isAutoRepeat) {
+        if (api.keys.isPrevPage(event) && !event.isAutoRepeat && !global.guideButtonPressed) {
             event.accepted = true;
             sfxToggle.play();
             if (currentGameIndex > 0)